add tests for NamespaceScopedCache.List()
